+++ Sources/PinLayout.swift
@@ -16,482 +16,976 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-/*
+
+import Foundation
+
 #if os(iOS) || os(tvOS)
-import UIKit
+    import UIKit
+    public typealias PEdgeInsets = UIEdgeInsets
 #else
-import AppKit
+    import AppKit
+    public typealias PEdgeInsets = NSEdgeInsets
 #endif
 
-/// PinLayout interface
-public protocol PinLayout {
+public class PinLayout<View: Layoutable> {
+    internal let view: View
+    internal let keepTransform: Bool
+
+    internal var _top: CGFloat?       // offset from superview's top edge
+    internal var _left: CGFloat?      // offset from superview's left edge
+    internal var _bottom: CGFloat?    // offset from superview's top edge
+    internal var _right: CGFloat?     // offset from superview's left edge
+    
+    internal var _hCenter: CGFloat?
+    internal var _vCenter: CGFloat?
+    
+    internal var width: CGFloat?
+    internal var minWidth: CGFloat?
+    internal var maxWidth: CGFloat?
+    internal var height: CGFloat?
+    internal var minHeight: CGFloat?
+    internal var maxHeight: CGFloat?
+
+    internal var adjustSizeType: AdjustSizeType?
+
+    internal var shouldKeepViewDimension: Bool {
+        return adjustSizeType == nil
+    }
+    
+    internal var marginTop: CGFloat?
+    internal var marginLeft: CGFloat?
+    internal var marginBottom: CGFloat?
+    internal var marginRight: CGFloat?
+    internal var shouldPinEdges = false
+    
+    internal var justify: HorizontalAlign?
+    internal var align: VerticalAlign?
+    
+    internal var _marginTop: CGFloat { return marginTop ?? 0 }
+    internal var _marginLeft: CGFloat { return marginLeft ?? 0 }
+    internal var _marginBottom: CGFloat { return marginBottom ?? 0 }
+    internal var _marginRight: CGFloat { return marginRight ?? 0 }
+    
+    internal var isLayouted = false
+
+    init(view: View, keepTransform: Bool) {
+        self.view = view
+        self.keepTransform = keepTransform
+
+        #if os(iOS) || os(tvOS)
+        Pin.initPinLayout()
+        #endif
+    }
+    
+    deinit {
+        if !isLayouted && Pin.logMissingLayoutCalls {
+            warn("PinLayout commands have been issued without calling the 'layout()' method to complete the layout. (These warnings can be disabled by setting Pin.logMissingLayoutCalls to false)")
+        }
+        apply()
+    }
 
     #if os(iOS) || os(tvOS)
-    /// Expose a compatible `safeAreaInsets` for all iOS releases (including iOS 7/8/9/10/...). Available only for >= iOS 9
-    var safeArea: UIEdgeInsets { get }
+    public var safeArea: PEdgeInsets {
+        if let view = view as? UIView {
+            if #available(iOS 11.0, tvOS 11.0, *) {
+                return view.safeAreaInsets
+            } else {
+                return view.pinlayoutComputeSafeAreaInsets()
+            }
+        } else {
+            return PEdgeInsets.zero
+        }
+    }
     #endif
 
     //
     // MARK: Layout using distances from superview’s edges
     //
-    @discardableResult func top() -> PinLayout
-    @discardableResult func top(_ value: CGFloat) -> PinLayout
-    @discardableResult func top(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func top(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func top(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    // top, left, bottom, right
+    //
 
-    @discardableResult func left() -> PinLayout
-    @discardableResult func left(_ value: CGFloat) -> PinLayout
-    @discardableResult func left(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func left(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func left(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    @discardableResult
+    public func top() -> PinLayout {
+        top({ return "top()" })
+        return self
+    }
 
-    @discardableResult func bottom() -> PinLayout
-    @discardableResult func bottom(_ value: CGFloat) -> PinLayout
-    @discardableResult func bottom(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func bottom(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func bottom(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    @discardableResult
+    public func top(_ value: CGFloat) -> PinLayout {
+        return top(value, { return "top(\(value))" })
+    }
 
-    @discardableResult func right() -> PinLayout
-    @discardableResult func right(_ value: CGFloat) -> PinLayout
-    @discardableResult func right(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func right(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func right(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    @discardableResult
+    public func top(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "top(\(percent.description))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setTop(percent.of(layoutSuperviewRect.height), context)
+        return self
+    }
 
-    @discardableResult func hCenter() -> PinLayout
-    @discardableResult func hCenter(_ value: CGFloat) -> PinLayout
-    @discardableResult func hCenter(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func top(_ insets: PEdgeInsets) -> PinLayout {
+        return top(insets.top, { return "top(\(insetsDescription(insets))" })
+    }
 
-    @discardableResult func vCenter() -> PinLayout
-    @discardableResult func vCenter(_ value: CGFloat) -> PinLayout
-    @discardableResult func vCenter(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func left() -> PinLayout {
+        return left({ return "left()" })
+    }
 
-    // RTL support
-    @discardableResult func start() -> PinLayout
-    @discardableResult func start(_ value: CGFloat) -> PinLayout
-    @discardableResult func start(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func start(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func start(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    @discardableResult
+    public func left(_ value: CGFloat) -> PinLayout {
+        return left(value, { return "left(\(value))" })
+    }
 
-    @discardableResult func end() -> PinLayout
-    @discardableResult func end(_ value: CGFloat) -> PinLayout
-    @discardableResult func end(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func end(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func end(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    @discardableResult
+    public func left(_ percent: Percent) -> PinLayout {
+        return left(percent, { return "left(\(percent.description))" })
+    }
+
+    @discardableResult
+    public func left(_ insets: PEdgeInsets) -> PinLayout {
+        return left(insets.left, { return "left(\(insetsDescription(insets))" })
+    }
+
+    @discardableResult
+    public func start() -> PinLayout {
+        func context() -> String { return "start()" }
+        return isLTR() ? left(context) : right(context)
+    }
+
+    @discardableResult
+    public func start(_ value: CGFloat) -> PinLayout {
+        func context() -> String { return "start(\(value))" }
+        return isLTR() ? left(value, context) : right(value, context)
+    }
+
+    @discardableResult
+    public func start(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "start(\(percent.description))" }
+        return isLTR() ? left(percent, context) : right(percent, context)
+    }
+
+    @discardableResult
+    public func start(_ insets: PEdgeInsets) -> PinLayout {
+        func context() -> String { return "start(\(insetsDescription(insets))" }
+        return isLTR() ? left(insets.left, context) : right(insets.right, context)
+    }
+    
+    @discardableResult
+    public func bottom() -> PinLayout {
+        return bottom({ return "bottom()" })
+    }
+
+    @discardableResult
+    public func bottom(_ value: CGFloat) -> PinLayout {
+        return bottom(value, { return "bottom(\(value))" })
+    }
+
+    @discardableResult
+    public func bottom(_ percent: Percent) -> PinLayout {
+        return bottom(percent, { return "bottom(\(percent.description))" })
+    }
+
+    @discardableResult
+    public func bottom(_ insets: PEdgeInsets) -> PinLayout {
+        return bottom(insets.bottom, { return "bottom(\(insetsDescription(insets))" })
+    }
+
+    @discardableResult
+    public func right() -> PinLayout {
+        return right({ return "right()" })
+    }
+
+    @discardableResult
+    public func right(_ value: CGFloat) -> PinLayout {
+        return right(value, { return "right(\(value))" })
+    }
+
+    @discardableResult
+    public func right(_ percent: Percent) -> PinLayout {
+        return right(percent, { return "right(\(percent.description))" })
+    }
+
+    @discardableResult
+    public func right(_ insets: PEdgeInsets) -> PinLayout {
+        return right(insets.right, { return "right(\(insetsDescription(insets))" })
+    }
+    
+    @discardableResult
+    public func end() -> PinLayout {
+        func context() -> String { return "end()" }
+        return isLTR() ? right(context) : left(context)
+    }
+
+    @discardableResult
+    public func end(_ value: CGFloat) -> PinLayout {
+        func context() -> String { return "end(\(value))" }
+        return isLTR() ? right(value, context) : left(value, context)
+    }
+
+    @discardableResult
+    public func end(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "end(\(percent.description))" }
+        return isLTR() ? right(percent, context) : left(percent, context)
+    }
+
+    @discardableResult
+    public func end(_ insets: PEdgeInsets) -> PinLayout {
+        func context() -> String { return "end(\(insetsDescription(insets))" }
+        return isLTR() ? right(insets.right, context) : left(insets.left, context)
+    }
+
+    @discardableResult
+    public func hCenter() -> PinLayout {
+        func context() -> String { return "hCenter()" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setHorizontalCenter(layoutSuperviewRect.width / 2, context)
+        return self
+    }
+
+    @discardableResult
+    public func hCenter(_ value: CGFloat) -> PinLayout {
+        func context() -> String { return "hCenter(\(value))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setHorizontalCenter((layoutSuperviewRect.width / 2) + value, context)
+        return self
+    }
+
+    @discardableResult
+    public func hCenter(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "hCenter(\(percent.description))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setHorizontalCenter((layoutSuperviewRect.width / 2) + percent.of(layoutSuperviewRect.width), context)
+        return self
+    }
+
+    @discardableResult
+    public func vCenter() -> PinLayout {
+        func context() -> String { return "vCenter()" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setVerticalCenter(layoutSuperviewRect.height / 2, context)
+        return self
+    }
+
+    @discardableResult
+    public func vCenter(_ value: CGFloat) -> PinLayout {
+        func context() -> String { return "vCenter(\(value))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setVerticalCenter((layoutSuperviewRect.height / 2) + value, context)
+        return self
+    }
+
+    @discardableResult
+    public func vCenter(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "vCenter(\(percent.description))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setVerticalCenter((layoutSuperviewRect.height / 2) + percent.of(layoutSuperviewRect.height), context)
+        return self
+    }
 
     // Pin multiple edges at once.
+
     /**
      Pin all edges on its superview's corresponding edges (top, bottom, left, right).
 
      Similar to calling `view.top().bottom().left().right()`
      */
-    @discardableResult func all() -> PinLayout
-    @discardableResult func all(_ value: CGFloat) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func all(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func all(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    @discardableResult
+    public func all() -> PinLayout {
+        top({ "all() top coordinate" })
+        bottom({ "all() bottom coordinate" })
+        right({ "all() right coordinate" })
+        left({ "all() left coordinate" })
+        return self
+    }
+
+    /**
+     Pin all edges on its superview's corresponding edges (top, bottom, left, right).
+
+     Similar to calling `view.top().bottom().left().right()`
+     */
+    @discardableResult
+    public func all(_ value: CGFloat) -> PinLayout {
+        top(value,  { "all(\(value)) top coordinate" })
+        bottom(value,  { "all(\(value)) bottom coordinate" })
+        left(value,  { "all(\(value)) left coordinate" })
+        right(value,  { "all(\(value)) right coordinate" })
+        return self
+    }
+
+    /**
+     Pin all edges on its superview's corresponding edges (top, bottom, left, right).
+
+     Similar to calling `view.top().bottom().left().right()`
+     */
+    @discardableResult
+    public func all(_ insets: PEdgeInsets) -> PinLayout {
+        top(insets.top,  { "all(\(insets)) top coordinate" })
+        bottom(insets.bottom,  { "all(\(insets)) bottom coordinate" })
+        left(insets.left,  { "all(\(insets)) left coordinate" })
+        right(insets.right,  { "all(\(insets)) right coordinate" })
+        return self
+    }
 
     /**
      Pin the left and right edges on its superview's corresponding edges.
 
      Similar to calling `view.left().right()`.
      */
-    @discardableResult func horizontally() -> PinLayout
-    @discardableResult func horizontally(_ value: CGFloat) -> PinLayout
-    @discardableResult func horizontally(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func horizontally(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func horizontally(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    @discardableResult
+    public func horizontally() -> PinLayout {
+        right({ "horizontally() right coordinate" })
+        left({ "horizontally() left coordinate" })
+        return self
+    }
+
+    /**
+     Pin the left and right edges on its superview's corresponding edges.
+
+     Similar to calling `view.left().right()`.
+     */
+    @discardableResult
+    public func horizontally(_ value: CGFloat) -> PinLayout {
+        left(value, { return "horizontally(\(value)) left coordinate" })
+        right(value, { return "horizontally(\(value)) right coordinate" })
+        return self
+    }
+
+    /**
+     Pin the left and right edges on its superview's corresponding edges.
+
+     Similar to calling `view.left().right()`.
+     */
+    @discardableResult
+    public func horizontally(_ percent: Percent) -> PinLayout {
+        left(percent, { return "horizontally(\(percent.description)) left coordinate" })
+        right(percent, { return "horizontally(\(percent.description)) right coordinate" })
+        return self
+    }
+
+    /**
+     Pin the left and right edges on its superview's corresponding edges.
+
+     Similar to calling `view.left().right()`.
+     */
+    @discardableResult
+    public func horizontally(_ insets: PEdgeInsets) -> PinLayout {
+        left(insets.left, { return "horizontally(\(insets)) left coordinate" })
+        right(insets.right, { return "horizontally(\(insets)) right coordinate" })
+        return self
+    }
 
     /**
      Pin the **top and bottom edges** on its superview's corresponding edges.
 
      Similar to calling `view.top().bottom()`.
      */
-    @discardableResult func vertically() -> PinLayout
-    @discardableResult func vertically(_ value: CGFloat) -> PinLayout
-    @discardableResult func vertically(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func vertically() -> PinLayout {
+        top({ "vertically() top coordinate" })
+        bottom({ "vertically() bottom coordinate" })
+        return self
+    }
+
+    /**
+     Pin the **top and bottom edges** on its superview's corresponding edges.
+
+     Similar to calling `view.top().bottom()`.
+     */
+    @discardableResult
+    public func vertically(_ value: CGFloat) -> PinLayout {
+        top(value, { return "vertically(\(value)) top coordinate" })
+        bottom(value, { return "vertically(\(value)) bottom coordinate" })
+        return self
+    }
+
+    /**
+     Pin the **top and bottom edges** on its superview's corresponding edges.
+
+     Similar to calling `view.top().bottom()`.
+     */
+    @discardableResult
+    public func vertically(_ percent: Percent) -> PinLayout {
+        top(percent, { return "vertically(\(percent.description)) top coordinate" })
+        bottom(percent, { return "vertically(\(percent.description)) bottom coordinate" })
+        return self
+    }
+
     /**
      Pin the **top and bottom edges** on its superview's corresponding edges.
      The UIEdgeInsets.top is used to pin the top edge and the UIEdgeInsets.bottom
      for the bottom edge.
      */
-    #if os(iOS) || os(tvOS)
-    @discardableResult func vertically(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func vertically(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    @discardableResult
+    public func vertically(_ insets: PEdgeInsets) -> PinLayout {
+        top(insets.top, { return "vertically(\(insets)) top coordinate" })
+        bottom(insets.bottom, { return "vertically(\(insets)) bottom coordinate" })
+        return self
+    }
 
     //
     // MARK: Layout using edges
     //
-    @discardableResult func top(to edge: VerticalEdge) -> PinLayout
-    @discardableResult func vCenter(to edge: VerticalEdge) -> PinLayout
-    @discardableResult func bottom(to edge: VerticalEdge) -> PinLayout
-    @discardableResult func left(to edge: HorizontalEdge) -> PinLayout
-    @discardableResult func hCenter(to edge: HorizontalEdge) -> PinLayout
-    @discardableResult func right(to edge: HorizontalEdge) -> PinLayout
-    // RTL support
-    @discardableResult func start(to edge: HorizontalEdge) -> PinLayout
-    @discardableResult func end(to edge: HorizontalEdge) -> PinLayout
+    // top, left, bottom, right
+    //
+
+    @discardableResult
+    public func top(to edge: VerticalEdge) -> PinLayout {
+        func context() -> String { return relativeEdgeContext(method: "top", edge: edge) }
+        if let coordinate = computeCoordinate(forEdge: edge, context) {
+            setTop(coordinate, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func vCenter(to edge: VerticalEdge) -> PinLayout {
+        func context() -> String { return relativeEdgeContext(method: "vCenter", edge: edge) }
+        if let coordinate = computeCoordinate(forEdge: edge, context) {
+            setVerticalCenter(coordinate, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func bottom(to edge: VerticalEdge) -> PinLayout {
+        func context() -> String { return relativeEdgeContext(method: "bottom", edge: edge) }
+        if let coordinate = computeCoordinate(forEdge: edge, context) {
+            setBottom(coordinate, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func left(to edge: HorizontalEdge) -> PinLayout {
+        func context() -> String { return relativeEdgeContext(method: "left", edge: edge) }
+        if let coordinate = computeCoordinate(forEdge: edge, context) {
+            setLeft(coordinate, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func hCenter(to edge: HorizontalEdge) -> PinLayout {
+        func context() -> String { return relativeEdgeContext(method: "hCenter", edge: edge) }
+        if let coordinate = computeCoordinate(forEdge: edge, context) {
+            setHorizontalCenter(coordinate, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func right(to edge: HorizontalEdge) -> PinLayout {
+        func context() -> String { return relativeEdgeContext(method: "right", edge: edge) }
+        if let coordinate = computeCoordinate(forEdge: edge, context) {
+            setRight(coordinate, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func start(to edge: HorizontalEdge) -> PinLayout {
+        func context() -> String { return relativeEdgeContext(method: "start", edge: edge) }
+        if let coordinate = computeCoordinate(forEdge: edge, context) {
+            setStart(coordinate, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func end(to edge: HorizontalEdge) -> PinLayout {
+        func context() -> String { return relativeEdgeContext(method: "end", edge: edge) }
+        if let coordinate = computeCoordinate(forEdge: edge, context) {
+            setEnd(coordinate, context)
+        }
+        return self
+    }
 
     //
     // MARK: Layout using anchors
     //
-    @discardableResult func topLeft(to anchor: Anchor) -> PinLayout
-    @discardableResult func topLeft() -> PinLayout
-    @discardableResult func topStart(to anchor: Anchor) -> PinLayout
-    @discardableResult func topStart() -> PinLayout
+    // topLeft, topCenter, topRight,
+    // centerLeft, center, centerRight,
+    // bottomLeft, bottomCenter, bottomRight,
+    //
 
-    @discardableResult func topCenter(to anchor: Anchor) -> PinLayout
-    @discardableResult func topCenter() -> PinLayout
+    @discardableResult
+    public func topLeft(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "topLeft", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setTopLeft(coordinatesList[0], context)
+        }
+        return self
+    }
 
-    @discardableResult func topRight(to anchor: Anchor) -> PinLayout
-    @discardableResult func topRight() -> PinLayout
-    @discardableResult func topEnd(to anchor: Anchor) -> PinLayout
-    @discardableResult func topEnd() -> PinLayout
+    @discardableResult
+    public func topLeft() -> PinLayout {
+        return topLeft({ return "topLeft()" })
+    }
+    
+    fileprivate func topLeft(_ context: Context) -> PinLayout {
+        setTopLeft(CGPoint(x: 0, y: 0), { return "topLeft()" })
+        return self
+    }
 
-    @discardableResult func centerLeft(to anchor: Anchor) -> PinLayout
-    @discardableResult func centerLeft() -> PinLayout
-    @discardableResult func centerStart(to anchor: Anchor) -> PinLayout
-    @discardableResult func centerStart() -> PinLayout
+    @discardableResult
+    public func topStart(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "topStart", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setTop(coordinatesList[0].y, context)
+            setStart(coordinatesList[0].x, context)
+        }
+        return self
+    }
 
-    @discardableResult func center(to anchor: Anchor) -> PinLayout
-    @discardableResult func center() -> PinLayout
+    @discardableResult
+    public func topStart() -> PinLayout {
+        func context() -> String { return "topStart()" }
+        return isLTR() ? topLeft(context) : topRight(context)
+    }
 
-    @discardableResult func centerRight(to anchor: Anchor) -> PinLayout
-    @discardableResult func centerRight() -> PinLayout
-    @discardableResult func centerEnd(to anchor: Anchor) -> PinLayout
-    @discardableResult func centerEnd() -> PinLayout
+    @discardableResult
+    public func topCenter(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "topCenter", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setTopCenter(coordinatesList[0], context)
+        }
+        return self
+    }
 
-    @discardableResult func bottomLeft(to anchor: Anchor) -> PinLayout
-    @discardableResult func bottomLeft() -> PinLayout
-    @discardableResult func bottomStart(to anchor: Anchor) -> PinLayout
-    @discardableResult func bottomStart() -> PinLayout
+    @discardableResult
+    public func topCenter() -> PinLayout {
+        func context() -> String { return "topCenter()" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setTopCenter(CGPoint(x: layoutSuperviewRect.width / 2, y: 0), context)
+        return self
+    }
 
-    @discardableResult func bottomCenter(to anchor: Anchor) -> PinLayout
-    @discardableResult func bottomCenter() -> PinLayout
+    @discardableResult
+    public func topRight(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "topRight", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setTopRight(coordinatesList[0], context)
+        }
+        return self
+    }
 
-    @discardableResult func bottomRight(to anchor: Anchor) -> PinLayout
-    @discardableResult func bottomRight() -> PinLayout
-    @discardableResult func bottomEnd(to anchor: Anchor) -> PinLayout
-    @discardableResult func bottomEnd() -> PinLayout
+    @discardableResult
+    public func topRight() -> PinLayout {
+        return topRight({ return "topRight()" })
+    }
+    
+    fileprivate func topRight(_ context: Context) -> PinLayout {
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setTopRight(CGPoint(x: layoutSuperviewRect.width, y: 0), context)
+        return self
+    }
+
+    @discardableResult
+    public func topEnd(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "topEnd", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setTop(coordinatesList[0].y, context)
+            setEnd(coordinatesList[0].x, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func topEnd() -> PinLayout {
+        func context() -> String { return "topEnd()" }
+        return isLTR() ? topRight(context) : topLeft(context)
+    }
+
+    @discardableResult
+    public func centerLeft(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "centerLeft", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setCenterLeft(coordinatesList[0], context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func centerLeft() -> PinLayout {
+        return centerLeft({ return "centerLeft()" })
+    }
+    
+    fileprivate func centerLeft(_ context: Context) -> PinLayout {
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setCenterLeft(CGPoint(x: 0, y: layoutSuperviewRect.height / 2), context)
+        return self
+    }
+
+    @discardableResult
+    public func centerStart(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "centerStart", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setVerticalCenter(coordinatesList[0].y, context)
+            setStart(coordinatesList[0].x, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func centerStart() -> PinLayout {
+        func context() -> String { return "centerStart()" }
+        return isLTR() ? centerLeft(context) : centerRight(context)
+    }
+
+    @discardableResult
+    public func center(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "center", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setCenter(coordinatesList[0], context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func center() -> PinLayout {
+        func context() -> String { return "center()" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setCenter(CGPoint(x: layoutSuperviewRect.width / 2, y: layoutSuperviewRect.height / 2), context)
+        return self
+    }
+
+    @discardableResult
+    public func centerRight(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "centerRight", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setCenterRight(coordinatesList[0], context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func centerRight() -> PinLayout {
+        return centerRight({ return "centerRight()" })
+    }
+
+    fileprivate func centerRight(_ context: Context) -> PinLayout {
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setCenterRight(CGPoint(x: layoutSuperviewRect.width, y: layoutSuperviewRect.height / 2), context)
+        return self
+    }
+
+    @discardableResult
+    public func centerEnd(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "centerEnd", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setVerticalCenter(coordinatesList[0].y, context)
+            setEnd(coordinatesList[0].x, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func centerEnd() -> PinLayout {
+        func context() -> String { return "centerEnd()" }
+        return isLTR() ? centerRight(context) : centerLeft(context)
+    }
+
+    @discardableResult
+    public func bottomLeft(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "bottomLeft", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setBottomLeft(coordinatesList[0], context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func bottomLeft() -> PinLayout {
+        return bottomLeft({ return "bottomLeft()" })
+    }
+
+    fileprivate func bottomLeft(_ context: Context) -> PinLayout {
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setBottomLeft(CGPoint(x: 0, y: layoutSuperviewRect.height), context)
+        return self
+    }
+
+    @discardableResult
+    public func bottomStart(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "bottomStart", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setBottom(coordinatesList[0].y, context)
+            setStart(coordinatesList[0].x, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func bottomStart() -> PinLayout {
+        func context() -> String { return "bottomStart()" }
+        return isLTR() ? bottomLeft(context) : bottomRight(context)
+    }
+
+    @discardableResult
+    public func bottomCenter(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "bottomCenter", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setBottomCenter(coordinatesList[0], context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func bottomCenter() -> PinLayout {
+        func context() -> String { return "bottomCenter()" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setBottomCenter(CGPoint(x: layoutSuperviewRect.width / 2, y: layoutSuperviewRect.height), context)
+        return self
+    }
+
+    @discardableResult
+    public func bottomRight(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "bottomRight", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setBottomRight(coordinatesList[0], context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func bottomRight() -> PinLayout {
+        return bottomRight({ return "bottomRight()" })
+    }
+
+    fileprivate func bottomRight(_ context: Context) -> PinLayout {
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        setBottomRight(CGPoint(x: layoutSuperviewRect.width, y: layoutSuperviewRect.height), context)
+        return self
+    }
+
+    @discardableResult
+    public func bottomEnd(to anchor: Anchor) -> PinLayout {
+        func context() -> String { return relativeAnchorContext(method: "bottomEnd", anchor: anchor) }
+        if let coordinatesList = computeCoordinates(forAnchors: [anchor], context) {
+            setBottom(coordinatesList[0].y, context)
+            setEnd(coordinatesList[0].x, context)
+        }
+        return self
+    }
+
+    @discardableResult
+    public func bottomEnd() -> PinLayout {
+        func context() -> String { return "bottomEnd()" }
+        return isLTR() ? bottomRight(context) : bottomLeft(context)
+    }
 
     //
-    // MARK: Layout using relative positioning
+    // MARK: Width, height
     //
-    #if os(iOS) || os(tvOS)
-        @discardableResult func above(of: UIView) -> PinLayout
-        @discardableResult func above(of: [UIView]) -> PinLayout
-        @discardableResult func above(of: UIView, aligned: HorizontalAlign) -> PinLayout
-        @discardableResult func above(of: [UIView], aligned: HorizontalAlign) -> PinLayout
 
-        @discardableResult func below(of: UIView) -> PinLayout
-        @discardableResult func below(of: [UIView]) -> PinLayout
-        @discardableResult func below(of: UIView, aligned: HorizontalAlign) -> PinLayout
-        @discardableResult func below(of: [UIView], aligned: HorizontalAlign) -> PinLayout
+    @discardableResult
+    public func width(_ width: CGFloat) -> PinLayout {
+        return setWidth(width, { return "width(\(width))" })
+    }
 
-        @discardableResult func left(of: UIView) -> PinLayout
-        @discardableResult func left(of: [UIView]) -> PinLayout
-        @discardableResult func left(of: UIView, aligned: VerticalAlign) -> PinLayout
-        @discardableResult func left(of: [UIView], aligned: VerticalAlign) -> PinLayout
+    @discardableResult
+    public func width(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "width(\(percent.description))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        return setWidth(percent.of(layoutSuperviewRect.width), context)
+    }
 
-        @discardableResult func right(of: UIView) -> PinLayout
-        @discardableResult func right(of: [UIView]) -> PinLayout
-        @discardableResult func right(of: UIView, aligned: VerticalAlign) -> PinLayout
-        @discardableResult func right(of: [UIView], aligned: VerticalAlign) -> PinLayout
+    @discardableResult
+    public func width(of view: View) -> PinLayout {
+        let rect = view.getRect(keepTransform: keepTransform)
+        return setWidth(rect.width, { return "width(of: \(viewDescription(view)))" })
+    }
 
-        // RTL support
-        @discardableResult func before(of: UIView) -> PinLayout
-        @discardableResult func before(of: [UIView]) -> PinLayout
-        @discardableResult func before(of: UIView, aligned: VerticalAlign) -> PinLayout
-        @discardableResult func before(of: [UIView], aligned: VerticalAlign) -> PinLayout
-        @discardableResult func after(of: UIView) -> PinLayout
-        @discardableResult func after(of: [UIView]) -> PinLayout
-        @discardableResult func after(of: UIView, aligned: VerticalAlign) -> PinLayout
-        @discardableResult func after(of: [UIView], aligned: VerticalAlign) -> PinLayout
-    #elseif os(macOS)
-        @discardableResult func above(of: NSView) -> PinLayout
-        @discardableResult func above(of: [NSView]) -> PinLayout
-        @discardableResult func above(of: NSView, aligned: HorizontalAlign) -> PinLayout
-        @discardableResult func above(of: [NSView], aligned: HorizontalAlign) -> PinLayout
+    @discardableResult
+    public func minWidth(_ width: CGFloat) -> PinLayout {
+        setMinWidth(width, { return "minWidth(\(width))" })
+        return self
+    }
 
-        @discardableResult func below(of: NSView) -> PinLayout
-        @discardableResult func below(of: [NSView]) -> PinLayout
-        @discardableResult func below(of: NSView, aligned: HorizontalAlign) -> PinLayout
-        @discardableResult func below(of: [NSView], aligned: HorizontalAlign) -> PinLayout
+    @discardableResult
+    public func minWidth(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "minWidth(\(percent.description))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        return setMinWidth(percent.of(layoutSuperviewRect.width), context)
+    }
 
-        @discardableResult func left(of: NSView) -> PinLayout
-        @discardableResult func left(of: [NSView]) -> PinLayout
-        @discardableResult func left(of: NSView, aligned: VerticalAlign) -> PinLayout
-        @discardableResult func left(of: [NSView], aligned: VerticalAlign) -> PinLayout
+    @discardableResult
+    public func maxWidth(_ width: CGFloat) -> PinLayout {
+        setMaxWidth(width, { return "maxWidth(\(width))" })
+        return self
+    }
 
-        @discardableResult func right(of: NSView) -> PinLayout
-        @discardableResult func right(of: [NSView]) -> PinLayout
-        @discardableResult func right(of: NSView, aligned: VerticalAlign) -> PinLayout
-        @discardableResult func right(of: [NSView], aligned: VerticalAlign) -> PinLayout
+    @discardableResult
+    public func maxWidth(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "maxWidth(\(percent.description))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        return setMaxWidth(percent.of(layoutSuperviewRect.width), context)
+    }
 
-        // RTL support
-        @discardableResult func before(of: NSView) -> PinLayout
-        @discardableResult func before(of: [NSView]) -> PinLayout
-        @discardableResult func before(of: NSView, aligned: VerticalAlign) -> PinLayout
-        @discardableResult func before(of: [NSView], aligned: VerticalAlign) -> PinLayout
-        @discardableResult func after(of: NSView) -> PinLayout
-        @discardableResult func after(of: [NSView]) -> PinLayout
-        @discardableResult func after(of: NSView, aligned: VerticalAlign) -> PinLayout
-        @discardableResult func after(of: [NSView], aligned: VerticalAlign) -> PinLayout
-    #endif
+    @discardableResult
+    public func height(_ height: CGFloat) -> PinLayout {
+        return setHeight(height, { return "height(\(height))" })
+    }
+
+    @discardableResult
+    public func height(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "height(\(percent.description))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        return setHeight(percent.of(layoutSuperviewRect.height), context)
+    }
+
+    @discardableResult
+    public func height(of view: View) -> PinLayout {
+        let rect = view.getRect(keepTransform: keepTransform)
+        return setHeight(rect.height, { return "height(of: \(viewDescription(view)))" })
+    }
+
+    @discardableResult
+    public func minHeight(_ height: CGFloat) -> PinLayout {
+        setMinHeight(height, { return "minHeight(\(height))" })
+        return self
+    }
+
+    @discardableResult
+    public func minHeight(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "minHeight(\(percent.description))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        return setMinHeight(percent.of(layoutSuperviewRect.height), context)
+    }
+
+    @discardableResult
+    public func maxHeight(_ height: CGFloat) -> PinLayout {
+        setMaxHeight(height, { return "maxHeight(\(height))" })
+        return self
+    }
+
+    @discardableResult
+    public func maxHeight(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "maxHeight(\(percent.description))" }
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        return setMaxHeight(percent.of(layoutSuperviewRect.height), context)
+    }
 
     //
     // MARK: justify / align
     //
-    @discardableResult func justify(_: HorizontalAlign) -> PinLayout
-    @discardableResult func align(_: VerticalAlign) -> PinLayout
 
-    //
-    // MARK: Width, height and size
-    //
-    @discardableResult func width(_ width: CGFloat) -> PinLayout
-    @discardableResult func width(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func width(of view: UIView) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func width(of view: NSView) -> PinLayout
-    #endif
+    @discardableResult
+    public func justify(_ value: HorizontalAlign) -> PinLayout {
+        justify = value
+        return self
+    }
 
-    @discardableResult func minWidth(_ width: CGFloat) -> PinLayout
-    @discardableResult func minWidth(_ percent: Percent) -> PinLayout
-    @discardableResult func maxWidth(_ width: CGFloat) -> PinLayout
-    @discardableResult func maxWidth(_ percent: Percent) -> PinLayout
-
-    @discardableResult func height(_ height: CGFloat) -> PinLayout
-    @discardableResult func height(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func height(of view: UIView) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func height(of view: NSView) -> PinLayout
-    #endif
-
-    @discardableResult func minHeight(_ height: CGFloat) -> PinLayout
-    @discardableResult func minHeight(_ percent: Percent) -> PinLayout
-    @discardableResult func maxHeight(_ height: CGFloat) -> PinLayout
-    @discardableResult func maxHeight(_ percent: Percent) -> PinLayout
-
-    @discardableResult func size(_ size: CGSize) -> PinLayout
-    @discardableResult func size(_ sideLength: CGFloat) -> PinLayout
-    @discardableResult func size(_ percent: Percent) -> PinLayout
-    #if os(iOS) || os(tvOS)
-    @discardableResult func size(of view: UIView) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func size(of view: NSView) -> PinLayout
-    #endif
-
-    /**
-     Adjust the view's width & height to wrap all its subviews. The method also adjust subviews position to create a tight wrap.
-     */
-    @discardableResult func wrapContent() -> PinLayout
-    /**
-     Adjust the view's width & height to wrap all its subviews. The method also adds a padding around all subviews.
-
-     - Parameters:
-         - padding: Specify a padding value.
-     */
-    @discardableResult func wrapContent(padding: CGFloat) -> PinLayout
-    /**
-     Adjust the view's width & height to wrap all its subviews. The method also adds a padding around all subviews.
-
-     - Parameters:
-         - padding: Specify a padding using an UIEdgeInsets.
-     */
-    #if os(iOS) || os(tvOS)
-    @discardableResult func wrapContent(padding: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func wrapContent(padding: NSEdgeInsets) -> PinLayout
-    #endif
-
-    /**
-     Adjust the view's width AND/OR height to wrap all its subviews.
-
-     - Parameters:
-        - type: Specify the wrap type (.all, .horizontally, .vertically)
-     */
-    @discardableResult func wrapContent(_ type: WrapType) -> PinLayout
-    /**
-     Adjust the view's width AND/OR height to wrap all its subviews. The method also adds a padding around all subviews.
-
-     - Parameters:
-         - type: Specify the wrap type (.all, .horizontally, .vertically)
-         - padding: Specify a padding value.
-     */
-    @discardableResult func wrapContent(_ type: WrapType, padding: CGFloat) -> PinLayout
-    /**
-     Adjust the view's width AND/OR height to wrap all its subviews. The method also adds a padding around all subviews.
-
-     - Parameters:
-        - type: Specify the wrap type (.all, .horizontally, .vertically)
-        - padding: Specify a padding using an UIEdgeInsets.
-     */
-    #if os(iOS) || os(tvOS)
-    @discardableResult func wrapContent(_ type: WrapType, padding: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func wrapContent(_ type: WrapType, padding: NSEdgeInsets) -> PinLayout
-    #endif
-
-    /**
-     Set the view aspect ratio.
-
-     AspectRatio is applied only if a single dimension (either width or height) can be determined,
-     in that case the aspect ratio will be used to compute the other dimension.
-
-     * AspectRatio is defined as the ratio between the width and the height (width / height).
-     * An aspect ratio of 2 means the width is twice the size of the height.
-     * AspectRatio respects the min (minWidth/minHeight) and the max (maxWidth/maxHeight)
-     dimensions of an item.
-     */
-    @discardableResult func aspectRatio(_ ratio: CGFloat) -> PinLayout
-    /**
-     Set the view aspect ratio using another UIView's aspect ratio.
-
-     AspectRatio is applied only if a single dimension (either width or height) can be determined,
-     in that case the aspect ratio will be used to compute the other dimension.
-
-     * AspectRatio is defined as the ratio between the width and the height (width / height).
-     * AspectRatio respects the min (minWidth/minHeight) and the max (maxWidth/maxHeight)
-     dimensions of an item.
-     */
-    #if os(iOS) || os(tvOS)
-    @discardableResult func aspectRatio(of view: UIView) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func aspectRatio(of view: NSView) -> PinLayout
-    #endif
-
-    /**
-     If the layouted view is an UIImageView, this method will set the aspectRatio using
-     the UIImageView's image dimension.
-
-     For other types of views, this method as no impact.
-     */
-    #if os(iOS) || os(tvOS)
-    @discardableResult func aspectRatio() -> PinLayout
-    #endif
-
-    /**
-     The method adjust the view's size based on the view's `sizeThatFits()` method result.
-     PinLayout will adjust either the view's width or height based on the `fitType` parameter value.
-
-     Notes:
-     * If margin rules apply, margins will be applied when determining the reference dimension (width/height).
-     * The resulting size will always respect `minWidth` / `maxWidth` / `minHeight` / `maxHeight`.
-
-     - Parameter fitType: Identify the reference dimension (width / height) that will be used
-     to adjust the view's size:
-
-     .width: The method adjust the view's size based on the **reference width**.
-     * If properties related to the width have been pinned (e.g: width, left & right, margins, ...),
-     the **reference width will be determined by these properties**, if not the **current view's width**
-     will be used.
-     * The resulting width will always **match the reference width**.
-
-     .height: The method adjust the view's size based on the **reference height**.
-     * If properties related to the height have been pinned (e.g: height, top & bottom, margins, ...),
-     the **reference height will be determined by these properties**, if not the **current view's height**
-     will be used.
-     * The resulting height will always **match the reference height**.
-
-     .widthFlexible: Similar to `.width`, except that PinLayout won't constrain the resulting width to
-     match the reference width. The resulting width may be smaller or bigger depending on the view's
-     sizeThatFits(..) method result. For example a single line UILabel may returns a smaller width if its
-     string is smaller than the reference width.
-
-     .heightFlexible: Similar to `.height`, except that PinLayout won't constrain the resulting height to
-     match the reference height. The resulting height may be smaller or bigger depending on the view's
-     sizeThatFits(..) method result.
-
-     Examples:
-
-     ```
-     // Adjust the view's size based on a width of 100 pixels.
-     // The resulting width will always match the pinned property `width(100)`.
-     view.pin.width(100).sizeToFit(.width)
-
-     // Adjust the view's size based on view's current width.
-     // The resulting width will always match the view's original width.
-     // The resulting height will never be bigger than the specified `maxHeight`.
-     view.pin.sizeToFit(.width).maxHeight(100)
-
-     // Adjust the view's size based on 100% of the superview's height.
-     // The resulting height will always match the pinned property `height(100%)`.
-     view.pin.height(100%).sizeToFit(.height)
-
-     // Adjust the view's size based on view's current height.
-     // The resulting width will always match the view's original height.
-     view.pin.sizeToFit(.height)
-
-     // Since `.widthFlexible` has been specified, its possible that the resulting
-     // width will be smaller or bigger than 100 pixels, based of the label's sizeThatFits()
-     // method result.
-     label.pin.width(100).sizeToFit(.widthFlexible)
-     ```
-     */
-    #if os(iOS) || os(tvOS)
-    @discardableResult func sizeToFit(_ fitType: FitType) -> PinLayout
-    #endif
-
-    #if os(iOS) || os(tvOS)
-    @available(*, deprecated, message: "fitSize() is deprecated, please use sizeToFit(fitType: FitType)")
-    @discardableResult func fitSize() -> PinLayout
-    #endif
-
+    @discardableResult
+    public func align(_ value: VerticalAlign) -> PinLayout {
+        align = value
+        return self
+    }
+    
     //
     // MARK: Margins
     //
+
     /**
      Set the top margin.
      */
-    @discardableResult func marginTop(_ value: CGFloat) -> PinLayout
-    @discardableResult func marginTop(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func marginTop(_ value: CGFloat) -> PinLayout {
+        marginTop = value
+        return self
+    }
+
+    /**
+     Set the top margin.
+     */
+    @discardableResult
+    public func marginTop(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "marginTop(\(percent.description))" }
+        return marginTop(percent, context)
+    }
+
+    private func marginTop(_ percent: Percent, _ context: Context) -> Self {
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        marginTop = percent.of(layoutSuperviewRect.height)
+        return self
+    }
 
     /**
      Set the left margin.
      */
-    @discardableResult func marginLeft(_ value: CGFloat) -> PinLayout
-    @discardableResult func marginLeft(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func marginLeft(_ value: CGFloat) -> PinLayout {
+        marginLeft = value
+        return self
+    }
+
+    /**
+     Set the left margin.
+     */
+    @discardableResult
+    public func marginLeft(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "marginLeft(\(percent.description))" }
+        return marginLeft(percent, context)
+    }
+
+    private func marginLeft(_ percent: Percent, _ context: Context) -> Self {
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        marginLeft = percent.of(layoutSuperviewRect.width)
+        return self
+    }
 
     /**
      Set the bottom margin.
      */
-    @discardableResult func marginBottom(_ value: CGFloat) -> PinLayout
-    @discardableResult func marginBottom(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func marginBottom(_ value: CGFloat) -> PinLayout {
+        marginBottom = value
+        return self
+    }
+
+    /**
+     Set the bottom margin.
+     */
+    @discardableResult
+    public func marginBottom(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "marginBottom(\(percent.description))" }
+        return marginBottom(percent, context)
+    }
+
+    private func marginBottom(_ percent: Percent, _ context: Context) -> Self {
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        marginBottom = percent.of(layoutSuperviewRect.height)
+        return self
+    }
 
     /**
      Set the right margin.
      */
-    @discardableResult func marginRight(_ value: CGFloat) -> PinLayout
-    @discardableResult func marginRight(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func marginRight(_ value: CGFloat) -> PinLayout {
+        marginRight = value
+        return self
+    }
+
+    /**
+     Set the right margin.
+     */
+    @discardableResult
+    public func marginRight(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "marginRight(\(percent.description))" }
+        return marginRight(percent, context)
+    }
+
+    private func marginRight(_ percent: Percent, _ context: Context) -> Self {
+        guard let layoutSuperviewRect = layoutSuperviewRect(context) else { return self }
+        marginRight = percent.of(layoutSuperviewRect.width)
+        return self
+    }
 
     // RTL support
+
     /**
      Set the start margin.
 
@@ -499,8 +993,23 @@ public protocol PinLayout {
      * In LTR direction, start margin specify the **left** margin.
      * In RTL direction, start margin specify the **right** margin.
      */
-    @discardableResult func marginStart(_ value: CGFloat) -> PinLayout
-    @discardableResult func marginStart(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func marginStart(_ value: CGFloat) -> PinLayout {
+        return isLTR() ? marginLeft(value) : marginRight(value)
+    }
+
+    /**
+     Set the start margin.
+
+     Depends on the value of `Pin.layoutDirection(...)`:
+     * In LTR direction, start margin specify the **left** margin.
+     * In RTL direction, start margin specify the **right** margin.
+     */
+    @discardableResult
+    public func marginStart(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "marginStart(\(percent.description))" }
+        return isLTR() ? marginLeft(percent, context) : marginRight(percent, context)
+    }
 
     /**
      Set the end margin.
@@ -509,30 +1018,82 @@ public protocol PinLayout {
      * In LTR direction, end margin specify the **right** margin.
      * In RTL direction, end margin specify the **left** margin.
      */
-    @discardableResult func marginEnd(_ value: CGFloat) -> PinLayout
-    @discardableResult func marginEnd(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func marginEnd(_ value: CGFloat) -> PinLayout {
+        return isLTR() ? marginRight(value) : marginLeft(value)
+    }
+
+    /**
+     Set the end margin.
+
+     Depends on the value of `Pin.layoutDirection(...)`:
+     * In LTR direction, end margin specify the **right** margin.
+     * In RTL direction, end margin specify the **left** margin.
+     */
+    @discardableResult
+    public func marginEnd(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "marginEnd(\(percent.description))" }
+        return isLTR() ? marginRight(percent, context) : marginLeft(percent, context)
+    }
 
     /**
      Set the left, right, start and end margins to the specified value.
      */
-    @discardableResult func marginHorizontal(_ value: CGFloat) -> PinLayout
-    @discardableResult func marginHorizontal(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func marginHorizontal(_ value: CGFloat) -> PinLayout {
+        marginLeft = value
+        marginRight = value
+        return self
+    }
+
+    /**
+     Set the left, right, start and end margins to the specified value.
+     */
+    @discardableResult
+    public func marginHorizontal(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "marginHorizontal(\(percent.description))" }
+        return marginHorizontal(percent, context)
+    }
+
+    private func marginHorizontal(_ percent: Percent, _ context: Context) -> Self {
+        return marginLeft(percent, context).marginRight(percent, context)
+    }
 
     /**
      Set the top and bottom margins to the specified value.
      */
-    @discardableResult func marginVertical(_ value: CGFloat) -> PinLayout
-    @discardableResult func marginVertical(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func marginVertical(_ value: CGFloat) -> PinLayout {
+        marginTop = value
+        marginBottom = value
+        return self
+    }
+
+    /**
+     Set the top and bottom margins to the specified value.
+     */
+    @discardableResult
+    public func marginVertical(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "marginVertical(\(percent.description))" }
+        return marginVertical(percent, context)
+    }
+
+    private func marginVertical(_ percent: Percent, _ context: Context) -> Self {
+        return marginTop(percent, context).marginBottom(percent, context)
+    }
 
     /**
      Set all margins using UIEdgeInsets.
      This method is particularly useful to set all margins using iOS 11 `UIView.safeAreaInsets`.
      */
-    #if os(iOS) || os(tvOS)
-    @discardableResult func margin(_ insets: UIEdgeInsets) -> PinLayout
-    #elseif os(macOS)
-    @discardableResult func margin(_ insets: NSEdgeInsets) -> PinLayout
-    #endif
+    @discardableResult
+    public func margin(_ insets: PEdgeInsets) -> PinLayout {
+        marginTop = insets.top
+        marginBottom = insets.bottom
+        marginLeft = insets.left
+        marginRight = insets.right
+        return self
+    }
 
     /**
      Set margins using NSDirectionalEdgeInsets.
@@ -542,32 +1103,107 @@ public protocol PinLayout {
      */
     #if os(iOS) || os(tvOS)
     @available(tvOS 11.0, iOS 11.0, *)
-    @discardableResult func margin(_ directionalInsets: NSDirectionalEdgeInsets) -> PinLayout
+    @discardableResult
+    public func margin(_ directionalInsets: NSDirectionalEdgeInsets) -> PinLayout {
+        marginTop = directionalInsets.top
+        marginBottom = directionalInsets.bottom
+        marginStart(directionalInsets.leading)
+        marginEnd(directionalInsets.trailing)
+        return self
+    }
     #endif
 
     /**
      Set all margins to the specified value.
      */
-    @discardableResult func margin(_ value: CGFloat) -> PinLayout
-    @discardableResult func margin(_ percent: Percent) -> PinLayout
+    @discardableResult
+    public func margin(_ value: CGFloat) -> PinLayout {
+        marginTop = value
+        marginLeft = value
+        marginBottom = value
+        marginRight = value
+        return self
+    }
 
     /**
-     Set individually vertical margins (top, bottom) and horizontal margins (left, right, start, end).
+     Set all margins to the specified value.
      */
-    @discardableResult func margin(_ vertical: CGFloat, _ horizontal: CGFloat) -> PinLayout
-    @discardableResult func margin(_ vertical: Percent, _ horizontal: Percent) -> PinLayout
+    @discardableResult
+    public func margin(_ percent: Percent) -> PinLayout {
+        func context() -> String { return "margin(\(percent.description))" }
+        return marginTop(percent, context)
+            .marginLeft(percent, context)
+            .marginBottom(percent, context)
+            .marginRight(percent, context)
+    }
 
     /**
      Set individually top, horizontal margins and bottom margin.
      */
-    @discardableResult func margin(_ top: CGFloat, _ horizontal: CGFloat, _ bottom: CGFloat) -> PinLayout
-    @discardableResult func margin(_ top: Percent, _ horizontal: Percent, _ bottom: Percent) -> PinLayout
+    @discardableResult
+    public func margin(_ top: CGFloat, _ left: CGFloat, _ bottom: CGFloat, _ right: CGFloat) -> PinLayout {
+        marginTop = top
+        marginLeft = left
+        marginBottom = bottom
+        marginRight = right
+        return self
+    }
 
     /**
-     Set individually top, left, bottom and right margins.
+     Set individually top, horizontal margins and bottom margin.
      */
-    @discardableResult func margin(_ top: CGFloat, _ left: CGFloat, _ bottom: CGFloat, _ right: CGFloat) -> PinLayout
-    @discardableResult func margin(_ top: Percent, _ left: Percent, _ bottom: Percent, _ right: Percent) -> PinLayout
+    @discardableResult
+    public func margin(_ top: Percent, _ left: Percent, _ bottom: Percent, _ right: Percent) -> PinLayout {
+        func context() -> String {
+            return "margin(top: \(top.description), left: \(left.description), bottom: \(bottom.description), right: \(right.description)"
+        }
+        return marginTop(top, context)
+            .marginLeft(left, context)
+            .marginBottom(bottom, context)
+            .marginRight(right, context)
+    }
+
+    /**
+     Set individually vertical margins (top, bottom) and horizontal margins (left, right, start, end).
+     */
+    @discardableResult
+    public func margin(_ vertical: CGFloat, _ horizontal: CGFloat) -> PinLayout {
+        marginTop = vertical
+        marginLeft = horizontal
+        marginBottom = vertical
+        marginRight = horizontal
+        return self
+    }
+
+    /**
+     Set individually vertical margins (top, bottom) and horizontal margins (left, right, start, end).
+     */
+    @discardableResult
+    public func margin(_ vertical: Percent, _ horizontal: Percent) -> PinLayout {
+        func context() -> String { return "margin(vertical: \(vertical.description), horizontal: \(horizontal.description)"}
+        return marginVertical(vertical, context).marginHorizontal(horizontal, context)
+    }
+
+    /**
+     Set individually top, horizontal margins and bottom margin.
+     */
+    @discardableResult
+    public func margin(_ top: CGFloat, _ horizontal: CGFloat, _ bottom: CGFloat) -> PinLayout {
+        marginTop = top
+        marginLeft = horizontal
+        marginBottom = bottom
+        marginRight = horizontal
+        return self
+    }
+
+    /**
+     Set individually top, horizontal margins and bottom margin.
+     */
+    @discardableResult
+    public func margin(_ top: Percent, _ horizontal: Percent, _ bottom: Percent) -> PinLayout {
+        func context() -> String { return "margin(top: \(top.description), horizontal: \(horizontal.description), bottom: \(bottom.description)"}
+        return marginTop(top, context).marginHorizontal(horizontal, context).marginBottom(bottom, context)
+    }
 
     /// Normally if only either left or right has been specified, PinLayout will MOVE the view to apply left or right margins.
     /// This is also true even if the width has been set.
@@ -575,18 +1211,46 @@ public protocol PinLayout {
     /// moving the view.
     ///
     /// - Returns: PinLayout
-    @discardableResult func pinEdges() -> PinLayout
-
-    /**
-     The method will execute PinLayout commands immediately. This method is **required only if your
-     source codes should also work in Xcode Playgrounds**. Outside of playgrounds, PinLayout executes
-     this method implicitly, it is not necessary to call it.
-
-     Examples:
-     ```swift
-     view.pin.top(20).width(100).layout()
-     ```
-     */
-    func layout()
+    @discardableResult
+    public func pinEdges() -> PinLayout {
+        shouldPinEdges = true
+        return self
+    }
 }
- */
+
+//
+// MARK: Private methods
+//
+extension PinLayout {
+    internal func layoutSuperviewRect(_ context: Context) -> CGRect? {
+        if let superview = view.superview {
+            return superview.getRect(keepTransform: keepTransform)
+        } else {
+            // Disable this warning: Using XIB, layoutSubview() is called even before views have been
+            // added, and there is no way to modify that strange behaviour of UIKit.
+            //warnWontBeApplied("the view must be added as a sub-view before being layouted using this method.", context)
+            return nil
+        }
+    }
+    
+    internal func layoutSuperview(_ context: Context) -> View? {
+        if let superview = view.superview {
+            return superview as? View
+        } else {
+            // Disable this warning: Using XIB, layoutSubview() is called even before views have been
+            // added, and there is no way to modify that strange behaviour of UIKit.
+            //warnWontBeApplied("the view must be added as a sub-view before being layouted using this method.", context)
+            return nil
+        }
+    }
+
+    internal func referenceSuperview(_ referenceView: View, _ context: Context) -> View? {
+        if let superview = referenceView.superview {
+            return superview as? View
+        } else {
+            warnWontBeApplied("the reference view \(viewDescription(referenceView)) must be added as a sub-view before being used as a reference.", context)
+            return nil
+        }
+    }
+}
+
